requirements.txt
flask
flask-socketio
whitenoise
eventlet
python-socketio[client]
RPi.GPIO